removed type cast

+++ 14-binary_tree_balance.c
@@ -9,11 +9,11 @@
  */
 int binary_tree_balance(const binary_tree_t *tree)
 {
-	int _height_r, _height_l;
+	size_t _height_r, _height_l;
 
 	if (!tree)
 		return (0);
 	_height_r = tree->right ? 1 + binary_tree_balance(tree->right) : 0;
 	_height_l = tree->left ? 1 + binary_tree_balance(tree->left) : 0;
-	return ((int)(_height_l - _height_r));
+	return (_height_l - _height_r);
 }
